requirements.txt
google-cloud-aiplatform>=1.125.0
google-generativeai==0.5.2
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv==1.0.0
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
starlette>=0.46.2
google-cloud-storage>=3.0.0